feat: IDE-35-ProjectService 테스트를 위한 ContainerService 인터페이스 추출
